# By Scott Walker (15) and others
# Via Scott Walker
* commit '1e1b6b02677709a916c6ffc6367f00e8bfaf318b':
  Added the CMake file
  merged Todd's changes, cpuid now uses 64 bit, added function to msr_core, fixed up init
  fixed a small bug in core.h
  fixed up init control flow, added losts of helpful error messages
  fixed up init control flow, added lots of helpful error messages
  fixed up init control flow, added lots of helpful error messages
  fixed up init control flow and added helpful error messages
  test2
  test
  putting up new rapl and msub
  added many checks for errors, made code more concise, improved init control flow
  added new error messages, checks for group perms, fixed a DRAM power cap bug
  Added error messages, checks for permissions, fixed a DRAM bug
  Removed test file which should be ignored
  New File
  found 1 more register that was missed by %s
  fixed cpuid to use uint64 instead of uint32
  fixed cpuid.c to use 64 bit registers

+++ include/cpuid.h
@@ -28,7 +28,7 @@ bool cpuid_therm_stat_powerlimit();
 bool cpuid_therm_stat_readout();
 bool cpuid_therm_interrupt_powerlimit();
 bool cpuid_pkg_therm_Stat_AND_Interrupt();
-int32_t cpuid_maxleaf();
+uint64_t cpuid_maxleaf();
 void cpuid_printVendorID();
 int cpuid_pkg_maxPhysicalProcessorCores();
 int cpuid_pkg_maxLogicalProcessors();
@@ -39,4 +39,4 @@ int cpuid_width_fixed_perf_counters();
 }
 #endif
 
-#endif // MSR_CHECKCPUID_H
+#endif // CHECKCPUID_H
